Moved grabber controls to their own subsystem and modified the Grabber command to use those subsystems.
Added grabber open / close commands to autonomous mode - to make the robot scary

Mapped the vertical and horiztonal camera servers to object in RobotMap.
Created new Joystick buttons to execute various camera commands.
Created a CameraControl subsystem with the following positional methods
- left
- right
- center
- up
- down
- level

Created Group Command and attached it to a button that combines CameraControl commands
to point the camera to a predefined postion.